requirements.txt
Flask==1.0.2
google-cloud-bigquery==1.7.0
beautifulsoup4==4.6.3
urllib3==1.24
requests==2.20.0
google-api-python-client
httplib2==0.11.3 
oauth2client==4.1.3
